Create test for isolated pawn
Create a test based on the movetext in #150 to check if the heuristic picture has been correctly calculated for the isolated pawn evaluation.
At the moment, this test fails because the heuristic has a value of 1 (which means white advantage) while its value should be -1 (which means black advantage).
The balance value of the isolated pawn is calculated inversely.